server/requirements.txt
fastapi 
uvicorn
python-multipart


langchain
langchain-community
langchain-core 
langchain-groq

pinecone

langchain-google-genai

tqdm
PyPDF
python-dotenv
pydantic
requests

loguru
